% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_duplicate.R
\name{index_duplicate}
\alias{index_duplicate}
\title{Duplicate strings with indices}
\usage{
index_duplicate(x, n)
}
\arguments{
\item{x}{A vector of strings}

\item{n}{A vector of numbers of times each string is to be repeated}
}
\value{
A vector of strings
}
\description{
Repeat strings in a vector and append an index to each repeat.
}
