% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_individual_data.R
\name{read_individual_data}
\alias{read_individual_data}
\title{Read individual data}
\usage{
read_individual_data(root)
}
\arguments{
\item{root}{Path to the simulation folder}
}
\value{
A tibble
}
\description{
This function is a wrapper around \code{read_data} tailored at reading the
output file "individuals.dat" into a tibble with one row per individual and
individual attributes (e.g. deme, patch and traits) in columns, properly
renamed.
}
\examples{

root <- system.file("extdata", "sim-example", package = "brachypoder")
read_individual_data(root)

}
\seealso{
\code{read_data}
}
