% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parameters.R
\name{read_parameters}
\alias{read_parameters}
\title{Read simulation parameters}
\usage{
read_parameters(root)
}
\arguments{
\item{root}{Path to the simulation folder}
}
\value{
A named list of parameter values
}
\description{
Read the parameter file in a simulation folder
}
\details{
The function looks for the file "paramlog.txt" in the simulation folder
}
\examples{

root <- system.file("extdata", "sim-example", package = "brachypoder")
read_parameters(root)

}
