% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_patch_size_data.R
\name{read_patch_size_data}
\alias{read_patch_size_data}
\title{Read patch size data}
\usage{
read_patch_size_data(root)
}
\arguments{
\item{root}{Path to the simulation folder}
}
\value{
A tibble
}
\description{
This function is a wrapper around \code{read_data} tailored at reading the
output file "patchsizes.dat" into a tibble with one row per deme and per patch.
}
\examples{

root <- system.file("extdata", "sim-example", package = "brachypoder")
read_patch_size_data(root)

}
\seealso{
\code{read_data}
}
