% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mean_trait_data.R
\name{read_trait_mean_data}
\alias{read_trait_mean_data}
\title{Read mean trait data}
\usage{
read_trait_mean_data(root)
}
\arguments{
\item{root}{Path to the simulation folder}
}
\value{
A tibble
}
\description{
This function is a wrapper around \code{read_data} tailored at reading the
output file "traitmeans.dat" into a tibble with one row per deme and per patch
and one column per trait.
}
\examples{

root <- system.file("extdata", "sim-example", package = "brachypoder")
read_trait_mean_data(root)

}
\seealso{
\code{read_data}
}
