% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2tbl.R
\name{sim2tbl}
\alias{sim2tbl}
\title{Turn a data vector into a tibble}
\usage{
sim2tbl(x, ncol = 1)
}
\arguments{
\item{x}{Vector of data values}

\item{ncol}{If integer, number of columns by which to split the vector.
Use a negative value to duplicate the vector into a longer one-column tibble instead of
splitting it. If integer vector, then each element in the \code{x} will be
duplicated a number of times defined by its corresponding (positive) entry in
\code{ncol}. So in this case, \code{ncol} and \code{x} must have the same length.}
}
\value{
A tibble containing the simulation data
}
\description{
Reshapes a simulation data vector into a matrix and converts it into a tibble.
}
\examples{

sim2tbl(1:3)
sim2tbl(1:3, ncol = 3)
sim2tbl(1:3, ncol = -2)
sim2tbl(1:3, ncol = 1:3)

}
\seealso{
\code{read_data}
}
